refactor: Rename SafeTxWatcher to SimplePoller
